+++ src/visuals/fermion_renderers/render_fermion_spin_dial.hlsl
@@ -9,9 +9,13 @@
 
 namespace FermionRendering
 {
+    /// Implementation of fermion field rendering by mapping the spin state onto vector field dials.
+    /// * Functions may read directly from and/or write directly to the simulation's lattice buffers and global values.
     namespace RenderFermionSpinDial
     {
-        half4 get_fermion_spin_dial_color_at_position(float3 position, uint field_index, half granularity, half length_scale)
+        // Get the color at a given position to render the spin state of the fermion field as a vector dial.
+        // • Reads directly from the simulation's lattice buffers
+        half4 get_fermion_spin_dial_color_at_position(half3 position, uint field_index, half granularity, half length_scale)
         {
             if (!SimulationDataOps::is_fermion_field_active(field_index)) return half4(0, 0, 0, 0);
             half3 rounded_position;
